[#77] fix: ui 수정
